Add a `Copy-BazecorBackups` cmdlet and fix encoding when copying files.
The new cmdlet (written using Copilot) will let me pass in the path to
the folder containing the backup json files exported from Bazecor and an
optional timestamp and then it will use the previous cmdlet I wrote to
sanitize each file and copy it into the repo, and then it commits the
changes from each backup file with a message including the timestamp
from the backup json file.
I changed the `Remove-NeuronIdFromBackupJson` cmdlet to use `Out-File`
with the `-Encoding` parameter in place of the redirection operator
(`>`) to ensure that the json files are saved with utf-8 encoding rather
than the PowerShell default of utf-16. This is necessary because git
will treat utf-16 files as being binary files and not show diffs.