Add some more cmdlets for formatting data from backups to be readable.
Now there's `Sync-BazecorBackups`, the renamed version of
`Copy-BazecorBackups`, that does what it did before, but also if there
was any change in the backup file it calls `Save-DefyLayouts`, which in
turn calls `Format-DefyKeyMap`, `Format-DefyColorMap`, and
`Format-DefyPalette` and saves the results into text files.
Most of this code was generated using Copilot. The color map was a bit
trickier for it and so that one I mostly figured out on my own, but with
some help from Copilot giving me a place to start from.